refactor(frontend): extract MermaidDiagram from page.tsx into components/
P1.b — first structural extract from the 5,400-line page.tsx. The
Mermaid renderer is self-contained (own useState, useRef, useEffect;
no coupling to the parent's state cascade) so moving it is a zero-risk
behavior-preserving change.

app/components/ is established as the home for subsequent component
extractions (Admin, System, Library, Agents panels).

Verified: frontend 43/43, production build succeeds, Mermaid diagrams
still render in the README preview path.
